Cleanup code, implement UInt32 class
Cleaned up the unnecessary parts of the codebase and added a UInt32 class which automatically handles the conversion to and from ctypes.uint32 without the need to manually cast after mathematical operations. This greatly streamlines the code and should help to reduce the amount of bugs from forgetting to or incorrectly cast from Python ints.